Fix syntax

+++ geoutils/spatial_tools.py
@@ -264,7 +264,7 @@ height2 and width2 are set based on reference's resolution and the maximum exten
             dst_bounds=dst_bounds,
             dst_res=reference_raster.res,
             dst_crs=reference_raster.crs,
-            dtype=reference_raster.data.dtype,
+            dst_dtype=reference_raster.data.dtype,
             dst_nodata=reference_raster.nodata,
             silent=True,
         )
